Split out parts of the test framework to make it easier for plugins to initialize tests
