+++ spec/spec_stubs.rb
@@ -4,7 +4,6 @@ class TestLog
   attr_reader :data
 
   def write(data)
-    puts "data=#{data}"
     @data = data
   end
 end
